Core is obese. Extract static fns related to scrolling
